fixed decimals on medical/pharmacy

+++ app/femr/ui/models/medical/CreateViewModelGet.java
@@ -10,14 +10,14 @@ public class CreateViewModelGet {
     private int age;
     private String sex;
     //begin vitals
-    private Float bloodPressureSystolic;
-    private Float bloodPressureDiastolic;
-    private Float heartRate;
+    private Integer bloodPressureSystolic;
+    private Integer bloodPressureDiastolic;
+    private Integer heartRate;
     private Float temperature;
-    private Float respiratoryRate;
+    private Integer respiratoryRate;
     private Float oxygenSaturation;
-    private Float heightFeet;
-    private Float heightInches;
+    private Integer heightFeet;
+    private Integer heightInches;
     private Float weight;
     //begin encounter
     private String chiefComplaint;
@@ -95,27 +95,27 @@ public class CreateViewModelGet {
         this.sex = sex;
     }
 
-    public Float getBloodPressureSystolic() {
+    public Integer getBloodPressureSystolic() {
         return bloodPressureSystolic;
     }
 
-    public void setBloodPressureSystolic(Float bloodPressureSystolic) {
+    public void setBloodPressureSystolic(Integer bloodPressureSystolic) {
         this.bloodPressureSystolic = bloodPressureSystolic;
     }
 
-    public Float getBloodPressureDiastolic() {
+    public Integer getBloodPressureDiastolic() {
         return bloodPressureDiastolic;
     }
 
-    public void setBloodPressureDiastolic(Float bloodPressureDiastolic) {
+    public void setBloodPressureDiastolic(Integer bloodPressureDiastolic) {
         this.bloodPressureDiastolic = bloodPressureDiastolic;
     }
 
-    public Float getHeartRate() {
+    public Integer getHeartRate() {
         return heartRate;
     }
 
-    public void setHeartRate(Float heartRate) {
+    public void setHeartRate(Integer heartRate) {
         this.heartRate = heartRate;
     }
 
@@ -135,27 +135,27 @@ public class CreateViewModelGet {
         this.oxygenSaturation = oxygenSaturation;
     }
 
-    public Float getRespiratoryRate() {
+    public Integer getRespiratoryRate() {
         return respiratoryRate;
     }
 
-    public void setRespiratoryRate(Float respiratoryRate) {
+    public void setRespiratoryRate(Integer respiratoryRate) {
         this.respiratoryRate = respiratoryRate;
     }
 
-    public Float getHeightFeet() {
+    public Integer getHeightFeet() {
         return heightFeet;
     }
 
-    public void setHeightFeet(Float heightFeet) {
+    public void setHeightFeet(Integer heightFeet) {
         this.heightFeet = heightFeet;
     }
 
-    public Float getHeightInches() {
+    public Integer getHeightInches() {
         return heightInches;
     }
 
-    public void setHeightInches(Float heightInches) {
+    public void setHeightInches(Integer heightInches) {
         this.heightInches = heightInches;
     }
